fix(#8) : 구조 및 코드 정리
