+++ app/src/main/java/com/example/ecarchargeinfo/main/presentation/binding/BindingAdapter.kt
@@ -1,12 +1,8 @@
-package com.example.ecarchargeinfo.binding
+package com.example.ecarchargeinfo.main.presentation.binding
 
 import android.view.View
-import androidx.databinding.BindingAdapter
 import androidx.lifecycle.LiveData
 import com.example.ecarchargeinfo.R
-import com.google.android.gms.maps.GoogleMap
-import com.google.android.gms.maps.OnMapReadyCallback
-import com.google.android.material.slider.RangeSlider
 
 object BindingAdapter {
