+++ app/src/main/java/com/example/ecarchargeinfo/main/presentation/helper/PermissionHelper.kt
@@ -1,0 +1,34 @@
+package com.example.ecarchargeinfo.main.presentation.helper
+
+import android.Manifest
+import android.app.Activity
+import android.content.pm.PackageManager
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
+import com.example.ecarchargeinfo.main.presentation.ui.MainActivity
+
+object PermissionHelper {
+    val REQUEST_PERMISSION = arrayOf(
+        Manifest.permission.ACCESS_FINE_LOCATION,
+        Manifest.permission.ACCESS_COARSE_LOCATION
+    )
+
+    fun checkPermission(activity: Activity) {
+        if (ContextCompat.checkSelfPermission(
+                activity,
+                Manifest.permission.ACCESS_FINE_LOCATION
+            ) == PackageManager.PERMISSION_DENIED ||
+            ContextCompat.checkSelfPermission(
+                activity,
+                Manifest.permission.ACCESS_COARSE_LOCATION
+            ) == PackageManager.PERMISSION_DENIED
+        ) {
+            ActivityCompat.requestPermissions(
+                activity,
+                REQUEST_PERMISSION,
+                MainActivity.LOCATION_PERMISSION
+            )
+        }
+    }
+
+}
