+++ app/src/main/java/com/example/ecarchargeinfo/map/domain/model/MapConstants.kt
@@ -1,0 +1,9 @@
+package com.example.ecarchargeinfo.map.domain.model
+
+object MapConstants {
+    val DEFAULT_LOCATION = arrayOf(
+        37.654601,
+        127.060530
+    )
+    const val DEFAULT_ZOOM = 15f
+}
